Fix share type rendering
I misunderstood how the field worked. It's a combination of a value enum
in the lower bits and a bitmask in the higher bits:
https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-srvs/6069f8c0-c93f-43a0-a5b4-7ed447eb4b84

In previous scan results, the "type_raw" fields are still correct, but
the "type" fields can have extra values. Most commonly, IPCs also have
erroneous DEVICE and PRINTQ labels (0x3 was interpreted as also implying
0x1 and 0x2). The cluster labels have a similar issue (but are rare in
practivce)